refactor(auth): split session.ts into 3 focused modules with barrel re-export
Split 904-line src/lib/auth/session.ts into:
- session-core.ts: session CRUD + cookie operations
- session-tokens.ts: token generation, validation, rotation
- session-remember-me.ts: remember-me token management
- session.ts: barrel re-export (backward compatible)
- index.ts: auth barrel for clean imports

Tests split into 4 focused files with 110 total tests
(+24 new tests for previously untested functions:
 createRememberMeToken, getRememberMeToken,
 deleteRememberMeToken, setAuthSessionCookie)

All 23 existing importers via @/lib/auth/session continue working.

Refs: #235

Cost-free: ✅
