+++ src/lib/auth/index.ts
@@ -1,0 +1,22 @@
+/**
+ * Auth Barrel Module
+ * Central re-export for all auth-related functions
+ */
+
+export {
+    createSession,
+    validateSession,
+    removeSession,
+    getSessionFromCookie,
+    setAuthSessionCookie,
+    generateSessionToken,
+    generateRememberMeToken,
+    validateSessionToken,
+    validateRememberMeToken,
+    refreshSessionToken,
+    createRememberMeToken,
+    getRememberMeToken,
+    deleteRememberMeToken,
+    storeRememberMeToken,
+    storeSessionToken,
+} from "./session"
